Replaced i16 with i32 in manualservice

+++ kraken/src/models/manual/mod.rs
@@ -60,7 +60,7 @@ pub struct ManualService {
 
     pub host: IpNetwork,
 
-    pub port: Option<i16>,
+    pub port: Option<i32>,
 
     /// The user which added the host
     pub user: ForeignModel<User>,
